Prevent timeout on long operations
By default Dokan has a 15s timeout for each request. If the application does
not finish the request in time the kernel driver will abort it with an error.
To prevent this the timeout is prolonged after 5s by one minute and, if
necessary, again as long as the request is pending.